# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TryNex Lifestyle API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products, admin]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProduct
      tags: [products, admin]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products, admin]
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Product deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: number
        discountPrice:
          type: number
        categoryId:
          type: integer
        categoryName:
          type: string
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        stock:
          type: integer
        featured:
          type: boolean
        rating:
          type: number
        reviewCount:
          type: integer
        customizable:
          type: boolean
        tags:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - slug
        - price
        - stock
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - products
        - total
        - page
        - limit
        - totalPages
    CreateProductRequest:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: number
        discountPrice:
          type: number
        categoryId:
          type: integer
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        stock:
          type: integer
        featured:
          type: boolean
        customizable:
          type: boolean
        tags:
          type: array
          items:
            type: string
      required:
        - name
        - slug
        - price
        - stock
    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: number
        discountPrice:
          type: number
        categoryId:
          type: integer
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        stock:
          type: integer
        featured:
          type: boolean
        customizable:
          type: boolean
        tags:
          type: array
          items:
            type: string
...