# lib/api-zod/src/generated/types/createProductRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TryNex Lifestyle API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProductRequest {
  name: string;
  slug: string;
  description?: string;
  price: number;
  discountPrice?: number;
  categoryId?: number;
  imageUrl?: string;
  images?: string[];
  sizes?: string[];
  colors?: string[];
  stock: number;
  featured?: boolean;
  customizable?: boolean;
  tags?: string[];
}